Move the wallet detail files into the wallet folder
