+++ lib/src/routes/app_route.dart
@@ -9,7 +9,7 @@ import 'package:flutter_xspend/src/sign_up/sign_up_view.dart';
 import 'package:flutter_xspend/src/new_transaction/new_transaction_view.dart';
 import 'package:flutter_xspend/src/clean_transaction/clean_transaction_view.dart';
 import 'package:flutter_xspend/src/profile/profile_view.dart';
-import 'package:flutter_xspend/src/wallet_details/wallet_detail_view.dart';
+import 'package:flutter_xspend/src/wallets/wallet_detail_view.dart';
 import 'package:flutter_xspend/src/budgets/new_budget_view.dart';
 import 'package:flutter_xspend/src/isar/isar_service.dart';
 import 'package:flutter_xspend/src/models/user.dart';
