+++ lib/src/wallets/wallet_view.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'package:flutter_xspend/src/constants/colors.dart';
-import 'package:flutter_xspend/src/wallets/wallet_header.dart';
-import 'package:flutter_xspend/src/wallets/wallet_line_chart.dart';
-import 'package:flutter_xspend/src/wallet_details/wallet_detail_view.dart';
+import 'wallet_header.dart';
+import 'wallet_line_chart.dart';
+import 'wallet_detail_view.dart';
 
 class WalletView extends StatelessWidget {
   const WalletView({super.key});
